Concludes the first version of shader compilation flags support
Allows the user to pass in their own flags, and sets the optimization level to O0 for debug builds and O3 for release builds, additionally the DEBUG macro is set to 1 for debug builds.
Currently no flag validation is done, such as in the case where multiple optimization flags are passed that would be conflicting, the same happens for macros, no validation of symbol redefinition is done.